implement armv7 platform, boot component, and initial HAL
This includes some of my contributions to platform/cav7, on top of the source tar of Renyu's
composite directory.  Renyu wrote much of this code for a paper, without the intention of
mainlining or maintaining it.  There is a lot of dead code.  There may be copied code
ithout attribution, or with license incompatibilities.

Renyu does not use git, so all of his work is in this commit.  Before mainlining the
arm branch, this commit will be squashed, and all of the dead code will be removed.
I am iteratively working to find and remove or attribute copied sections.