Do not pass arguments on stack using x86ism on arm when making syscall
On x86 regparm specifies to pass arguments on the stack, removing
a few loads/stores from the syscall hotpath.  This is not a valid
compilier extension on arm.

Future work: reoptimize this hotpath on arm.  register passing
arguments is not hard to do but is not the lowest hanging optimization